+++ src/routes/auth/wallet-signup.ts
@@ -1,7 +1,7 @@
 import { Response } from 'express'
 import { request } from '@shared/request'
 import { RequestExtended } from '@shared/types'
-import { asyncWrapper, selectAccount, verifySignature } from '@shared/helpers'
+import { asyncWrapper, getUserDataFromAccount, selectAccount, verifySignature } from '@shared/helpers'
 import { getAccountByWalletAddress,  insertAccountProviderWithUserAccount, trackUserSignUp } from '@shared/queries'
 import { setRefreshToken } from '@shared/cookies'
 import { AccountData, Session } from '@shared/types'
@@ -168,7 +168,8 @@ async function walletSignup(req: RequestExtended, res: Response): Promise<unknow
   const refresh_token = await setRefreshToken(res, account.id, useCookie)
   const jwt_token = createHasuraJwt(account)
   const jwt_expires_in = newJwtExpiry
-  const session: Session = { jwt_token, jwt_expires_in, user: account.user }
+
+  const session: Session = { jwt_token, jwt_expires_in, user: getUserDataFromAccount(account) }
   if (useCookie) session.refresh_token = refresh_token
 
   res.cookie('nonce', null) //empty nonce
